feat: add tools page
