+++ src/styles/index.css
@@ -1,5 +1,6 @@
 @import 'header.css';
 @import 'Home.css';
+@import 'Tools.css';
 @import 'About.css';
 @import 'Team.css';
 @import 'Footer.css';
